Address comments 3
