--- src/main/java/com/google/cloud/genomics/dataflow/utils/GcsTempPathOptions.java
@@ -1,44 +1,0 @@
-/*
- * Copyright (C) 2018 Google Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
- */
-package com.google.cloud.genomics.dataflow.utils;
-
-import org.apache.beam.sdk.options.Description;
-import org.apache.beam.sdk.options.Validation;
-import org.apache.beam.sdk.util.gcsfs.GcsPath;
-import com.google.common.base.Preconditions;
-import com.google.common.base.Strings;
-
-/**
- * A common options class for all pipelines that write into a temporary GCS files.
- */
-public interface GcsTempPathOptions extends GenomicsOptions {
-
-  public static class Methods {
-    public static void validateOptions(GcsTempPathOptions options) {
-      try {
-        // Check that we can parse the path.
-        GcsPath valid = GcsPath.fromUri(options.getTempPath());
-        // GcsPath allows for empty bucket, but that doesn't make for a good temp path.
-        Preconditions.checkArgument(!Strings.isNullOrEmpty(valid.getBucket()), "Bucket must be specified");
-      } catch (Exception x) {
-        Preconditions.checkState(false, "temp path must be a valid Google Cloud Storage URL (starting with gs://)");
-      }
-    }
-  }
-
-  @Validation.Required
-  @Description("Google Cloud Storage path to write temp files, if applicable.")
-  String getTempPath();
-  void setTempPath(String tempPath);
-}
